refactor(build-logic): move Android Compose configuration to android package
Refactored Android Compose setup for better structure and clarity:
- Moved `configureAndroidCompose` to the `android` package.
- Updated `AndroidApplicationComposeConventionPlugin` and `AndroidLibraryComposeConventionPlugin` to use the new location.
- Added new dependencies to `libs.versions.toml` for Compose-related libraries, including `androidx-compose-ui` and `kotlinCompiler`.
- Introduced custom compiler metrics and reports setup within `ComposeCompilerGradlePluginExtension`.
