+++ build-logic/convention/src/main/kotlin/AndroidApplicationComposeConventionPlugin.kt
@@ -1,5 +1,5 @@
 import com.android.build.api.dsl.ApplicationExtension
-import com.xbot.convention.configureAndroidCompose
+import com.xbot.convention.android.configureAndroidCompose
 import org.gradle.api.Plugin
 import org.gradle.api.Project
 import org.gradle.kotlin.dsl.getByType
