+++ build-logic/convention/src/main/kotlin/com/xbot/convention/android/AndroidCompose.kt
@@ -1,8 +1,15 @@
-@file:Suppress("UnstableApiUsage")
+/*
+ * Created by AnyGogin31 on 10.11.2024
+ */
 
-package com.xbot.convention
+package com.xbot.convention.android
 
 import com.android.build.api.dsl.CommonExtension
+import com.xbot.convention.extensions.debugImplementation
+import com.xbot.convention.extensions.getLibrary
+import com.xbot.convention.extensions.getPlugin
+import com.xbot.convention.extensions.getVersion
+import com.xbot.convention.extensions.implementation
 import com.xbot.convention.extensions.libs
 import org.gradle.api.Project
 import org.gradle.api.provider.Provider
@@ -11,34 +18,35 @@ import org.gradle.kotlin.dsl.configure
 import org.gradle.kotlin.dsl.dependencies
 import org.jetbrains.kotlin.compose.compiler.gradle.ComposeCompilerGradlePluginExtension
 
-/**
- * Configure Compose-specific options
- */
 internal fun Project.configureAndroidCompose(
     commonExtension: CommonExtension<*, *, *, *, *, *>,
 ) {
+    pluginManager.apply(libs.getPlugin("compose-compiler").get().pluginId)
+
     commonExtension.apply {
         buildFeatures {
             compose = true
         }
 
-        dependencies {
-            val bom = libs.findLibrary("androidx-compose-bom").get()
-            add("implementation", platform(bom))
-            add("androidTestImplementation", platform(bom))
-            add("implementation", libs.findLibrary("androidx-compose-ui-tooling-preview").get())
-            add("debugImplementation", libs.findLibrary("androidx-compose-ui-tooling").get())
+        composeOptions {
+            kotlinCompilerExtensionVersion = libs.getVersion("kotlinCompiler").toString()
         }
 
-        testOptions {
-            unitTests {
-                // For Robolectric
-                isIncludeAndroidResources = true
-            }
+        dependencies {
+            implementation(platform(libs.getLibrary("androidx-compose-bom")))
+
+            implementation(libs.getLibrary("androidx-compose-ui"))
+            implementation(libs.getLibrary("androidx-compose-ui-util"))
+            implementation(libs.getLibrary("androidx-compose-ui-tooling-preview"))
+
+            debugImplementation(libs.getLibrary("androidx-compose-ui-tooling"))
+
+            implementation(libs.getLibrary("androidx-compose-material"))
+            implementation(libs.getLibrary("androidx-compose-material3"))
         }
     }
 
-    //Command to run: ./gradlew assembleRelease -PenableComposeCompilerReports=true -PenableComposeCompilerMetrics=true
+    // Moved without changes
     extensions.configure<ComposeCompilerGradlePluginExtension> {
         fun Provider<String>.onlyIfTrue() = flatMap { provider { it.takeIf(String::toBoolean) } }
         fun Provider<*>.relativeToRootProject(dir: String) = flatMap {
