+++ gradle/libs.versions.toml
@@ -47,6 +47,7 @@ hiltExt = "1.2.0"
 jacoco = "0.8.7"
 junit4 = "4.13.2"
 kotlin = "2.0.10"
+kotlinCompiler = "1.5.15"
 kotlinxCollections = "0.3.7"
 kotlinxCoroutines = "1.9.0"
 kotlinxDatetime = "0.6.0"
@@ -89,6 +90,7 @@ androidx-compose-material3-adaptive-navigation = { group = "androidx.compose.mat
 androidx-compose-material-navigation = { group = "androidx.compose.material", name = "material-navigation", version = "1.8.0-alpha04" }
 androidx-compose-runtime = { group = "androidx.compose.runtime", name = "runtime" }
 androidx-compose-google-fonts = { group = "androidx.compose.ui", name = "ui-text-google-fonts" }
+androidx-compose-ui = { group = "androidx.compose.ui", name = "ui" }
 androidx-compose-ui-test = { group = "androidx.compose.ui", name = "ui-test-junit4" }
 androidx-compose-ui-testManifest = { group = "androidx.compose.ui", name = "ui-test-manifest" }
 androidx-compose-ui-tooling = { group = "androidx.compose.ui", name = "ui-tooling" }
